Separate optional and variant support from other STL containers #1207
