+++ setup.py
@@ -37,6 +37,7 @@ else:
         'include/pybind11/pytypes.h',
         'include/pybind11/stl.h',
         'include/pybind11/stl_bind.h',
+        'include/pybind11/utility.h',
     ]
